feat(assessments): delete assessment uses deletebutton
