+++ client/app/bundles/course/assessment/pages/AssessmentShow/AssessmentShowHeader.tsx
@@ -1,16 +1,21 @@
-import { Link } from 'react-router-dom';
-import { Create, Delete, Inventory } from '@mui/icons-material';
+import { useState } from 'react';
+import { Link, useNavigate } from 'react-router-dom';
+import { toast } from 'react-toastify';
+import { Create, Inventory } from '@mui/icons-material';
 import { Button, IconButton, Tooltip } from '@mui/material';
-import { AssessmentData } from 'types/course/assessment/assessments';
+import {
+  AssessmentData,
+  AssessmentDeleteResult,
+} from 'types/course/assessment/assessments';
 
+import DeleteButton from 'lib/components/core/buttons/DeleteButton';
+import { PromptText } from 'lib/components/core/dialogs/Prompt';
 import PageHeader from 'lib/components/navigation/PageHeader';
-import useToggle from 'lib/hooks/useToggle';
 import useTranslation from 'lib/hooks/useTranslation';
 
+import { deleteAssessment } from '../../actions';
 import translations from '../../translations';
 import { ACTION_LABELS } from '../AssessmentsIndex/ActionButtons';
-
-import DeleteAssessmentPrompt from './prompts/DeleteAssessmentPrompt';
 
 interface AssessmentShowHeaderProps {
   with: AssessmentData;
@@ -21,68 +26,87 @@ const AssessmentShowHeader = (
 ): JSX.Element => {
   const { with: assessment } = props;
   const { t } = useTranslation();
-  const [deleting, toggleDeleting] = useToggle();
+  const [deleting, setDeleting] = useState(false);
+  const navigate = useNavigate();
+
+  const handleDelete = (): Promise<void> => {
+    const deleteUrl = assessment.deleteUrl;
+    if (!deleteUrl)
+      throw new Error(
+        `Delete URL for assessment '${assessment.title}' is ${deleteUrl}.`,
+      );
+
+    setDeleting(true);
+
+    return toast
+      .promise(deleteAssessment(deleteUrl), {
+        pending: t(translations.deletingAssessment),
+        success: t(translations.assessmentDeleted),
+        error: {
+          render: ({ data }) => {
+            const error = (data as Error)?.message;
+            return error || t(translations.errorDeletingAssessment);
+          },
+        },
+      })
+      .then((data: AssessmentDeleteResult) => navigate(data.redirect))
+      .catch(() => setDeleting(false));
+  };
 
   return (
-    <>
-      <PageHeader returnLink={assessment.indexUrl} title={assessment.title}>
-        {assessment.deleteUrl && (
-          <Tooltip disableInteractive title={t(translations.deleteAssessment)}>
-            <IconButton
-              aria-label={t(translations.deleteAssessment)}
-              color="error"
-              onClick={toggleDeleting}
-            >
-              <Delete />
-            </IconButton>
-          </Tooltip>
-        )}
+    <PageHeader returnLink={assessment.indexUrl} title={assessment.title}>
+      {assessment.deleteUrl && (
+        <DeleteButton
+          aria-label={t(translations.deleteAssessment)}
+          confirmLabel={t(translations.deleteAssessment)}
+          disabled={deleting}
+          onClick={handleDelete}
+          title={t(translations.sureDeletingAssessment)}
+        >
+          <PromptText>{t(translations.deletingThisAssessment)}</PromptText>
+          <PromptText className="italic">{assessment.title}</PromptText>
+          <PromptText>{t(translations.deleteAssessmentWarning)}</PromptText>
+        </DeleteButton>
+      )}
 
-        {assessment.editUrl && (
-          <Tooltip disableInteractive title={t(translations.editAssessment)}>
-            <Link to={assessment.editUrl}>
-              <IconButton
-                aria-label={t(translations.editAssessment)}
-                className="text-white"
-              >
-                <Create />
-              </IconButton>
-            </Link>
-          </Tooltip>
-        )}
-
-        {assessment.submissionsUrl && (
-          <Tooltip disableInteractive title={t(translations.submissions)}>
+      {assessment.editUrl && (
+        <Tooltip disableInteractive title={t(translations.editAssessment)}>
+          <Link to={assessment.editUrl}>
             <IconButton
-              aria-label={t(translations.submissions)}
+              aria-label={t(translations.editAssessment)}
               className="text-white"
-              // TODO: Change to react-router Link once SPA
-              href={assessment.submissionsUrl}
             >
-              <Inventory />
+              <Create />
             </IconButton>
-          </Tooltip>
-        )}
+          </Link>
+        </Tooltip>
+      )}
 
-        {assessment.actionButtonUrl && (
-          <Button
-            aria-label={t(ACTION_LABELS[assessment.status])}
-            className="ml-4 bg-white"
+      {assessment.submissionsUrl && (
+        <Tooltip disableInteractive title={t(translations.submissions)}>
+          <IconButton
+            aria-label={t(translations.submissions)}
+            className="text-white"
             // TODO: Change to react-router Link once SPA
-            href={assessment.actionButtonUrl}
-            variant="outlined"
+            href={assessment.submissionsUrl}
           >
-            {t(ACTION_LABELS[assessment.status])}
-          </Button>
-        )}
-      </PageHeader>
+            <Inventory />
+          </IconButton>
+        </Tooltip>
+      )}
 
-      <DeleteAssessmentPrompt
-        for={assessment}
-        onClose={toggleDeleting}
-        open={deleting}
-      />
-    </>
+      {assessment.actionButtonUrl && (
+        <Button
+          aria-label={t(ACTION_LABELS[assessment.status])}
+          className="ml-4 bg-white"
+          // TODO: Change to react-router Link once SPA
+          href={assessment.actionButtonUrl}
+          variant="outlined"
+        >
+          {t(ACTION_LABELS[assessment.status])}
+        </Button>
+      )}
+    </PageHeader>
   );
 };
